@@ -1,49 +1,34 @@
 ﻿using System.Globalization;
 using CsvHelper;
-using DxWorks.ScriptBee.Plugin.Api.HelperFunctions;
+using DxWorks.ScriptBee.Plugin.Api;
 
 namespace DxWorks.ScriptBee.Plugin.HelperFunctions.Csv;
 
 public class CsvHelperFunctions : IHelperFunctions
 {
-    private readonly HelperFunctionsSettings _helperFunctionsSettings;
-    private readonly IFileModelService _fileModelService;
-    private readonly IFileNameGenerator _fileNameGenerator;
-    private readonly IResultCollector _resultCollector;
+    private readonly IHelperFunctionsResultService _helperFunctionsResultService;
 
-    public CsvHelperFunctions(HelperFunctionsSettings helperFunctionsSettings, IFileModelService fileModelService,
-        IFileNameGenerator fileNameGenerator, IResultCollector resultCollector)
+    public CsvHelperFunctions(IHelperFunctionsResultService helperFunctionsResultService)
     {
-        _helperFunctionsSettings = helperFunctionsSettings;
-        _fileModelService = fileModelService;
-        _fileNameGenerator = fileNameGenerator;
-        _resultCollector = resultCollector;
+        _helperFunctionsResultService = helperFunctionsResultService;
     }
 
-    public async Task ExportCsvAsync<T>(string fileName, List<T> records)
+    public async Task ExportCsvAsync<T>(string fileName, IEnumerable<T> records,
+        CancellationToken cancellationToken = default)
     {
-        var outputCsvName = _fileNameGenerator.GenerateOutputFileName(_helperFunctionsSettings.ProjectId,
-            _helperFunctionsSettings.RunId, RunResult.FileType, fileName);
-
-        _resultCollector.Add(new RunResult(RunResult.FileType, outputCsvName));
-
         await using var stream = new MemoryStream();
 
         await using var writer = new StreamWriter(stream);
         await using var csv = new CsvWriter(writer, CultureInfo.InvariantCulture);
 
-        await csv.WriteRecordsAsync(records);
+        await csv.WriteRecordsAsync(records, cancellationToken);
 
-        await _fileModelService.UploadFileAsync(outputCsvName, stream);
+        await _helperFunctionsResultService.UploadResultAsync(fileName, RunResultDefaultTypes.FileType, stream,
+            cancellationToken);
     }
 
     public void ExportCsv(string fileName, List<object> records)
     {
-        var outputCsvName = _fileNameGenerator.GenerateOutputFileName(_helperFunctionsSettings.ProjectId,
-            _helperFunctionsSettings.RunId, RunResult.FileType, fileName);
-
-        _resultCollector.Add(new RunResult(RunResult.FileType, outputCsvName));
-
         using var stream = new MemoryStream();
 
         using var writer = new StreamWriter(stream);
@@ -51,6 +36,6 @@ public class CsvHelperFunctions : IHelperFunctions
 
         csv.WriteRecords(records);
 
-        _fileModelService.UploadFile(outputCsvName, stream);
+        _helperFunctionsResultService.UploadResult(fileName, RunResultDefaultTypes.FileType, stream);
     }
 }
