@@ -1,29 +1,28 @@
-﻿using DxWorks.ScriptBee.Plugin.Api.Model;
-using DxWorks.ScriptBee.Plugin.Api.ScriptGeneration;
-using DxWorks.ScriptBee.Plugin.Api.ScriptRunner;
+﻿using DxWorks.ScriptBee.Plugin.Api;
+using DxWorks.ScriptBee.Plugin.Api.Model;
 
 namespace DxWorks.ScriptBee.Plugin.ScriptRunner.Python;
 
 public class PythonScriptRunner : IScriptRunner
 {
     private readonly IHelperFunctionsFactory _helperFunctionsFactory;
-    private readonly IHelperFunctionsMapper _helperFunctionsMapper;
     private readonly IScriptGeneratorStrategy _scriptGeneratorStrategy;
 
     public PythonScriptRunner(IHelperFunctionsFactory helperFunctionsFactory,
-        IHelperFunctionsMapper helperFunctionsMapper, IScriptGeneratorStrategy scriptGeneratorStrategy)
+        IScriptGeneratorStrategy scriptGeneratorStrategy)
     {
         _helperFunctionsFactory = helperFunctionsFactory;
-        _helperFunctionsMapper = helperFunctionsMapper;
         _scriptGeneratorStrategy = scriptGeneratorStrategy;
     }
+
+    public string Language => "python";
 
     public async Task RunAsync(IProject project, string runId, string scriptContent,
         CancellationToken cancellationToken = default)
     {
         var pythonEngine = IronPython.Hosting.Python.CreateEngine();
 
-        var helperFunctions = _helperFunctionsFactory.Create(project.Id, runId);
+        var helperFunctionsContainer = _helperFunctionsFactory.Create(project.Id, runId);
 
         var dictionary = new Dictionary<string, object>
         {
@@ -32,7 +31,7 @@ public class PythonScriptRunner : IScriptRunner
             },
         };
 
-        foreach (var (functionName, delegateFunction) in _helperFunctionsMapper.GetFunctionsDictionary(helperFunctions))
+        foreach (var (functionName, delegateFunction) in helperFunctionsContainer.GetFunctionsDictionary())
         {
             dictionary.Add(functionName, delegateFunction);
         }
