@@ -1,8 +1,15 @@
+using System;
+using System.Collections.Generic;
+using System.Linq.Expressions;
+
 namespace ScriptBee.Plugin;
 
+// todo maybe combine with plugin service
 public interface IPluginRepository
 {
-    void RegisterPlugin<T>(object argument);
+    void RegisterPlugin(object plugin);
 
-    T? GetPlugin<T>(object argument);
+    T? GetPlugin<T>(Expression<Func<T, bool>> filter) where T : class;
+
+    IEnumerable<T> GetPlugins<T>(Expression<Func<T, bool>> filter) where T : class;
 }
