@@ -2,7 +2,7 @@
 using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
-using DxWorks.ScriptBee.Plugin.Api.ScriptGeneration;
+using DxWorks.ScriptBee.Plugin.Api;
 
 namespace ScriptBeeWebApp.Services;
 
@@ -11,6 +11,6 @@ public interface IGenerateScriptService
     IEnumerable<string> GetSupportedLanguages();
     IScriptGeneratorStrategy? GetGenerationStrategy(string scriptType);
 
-    Task<Stream> GenerateClassesZip(List<object> classes, IScriptGeneratorStrategy scriptGeneratorStrategy,
+    Task<Stream> GenerateClassesZip(IEnumerable<object> classes, IScriptGeneratorStrategy scriptGeneratorStrategy,
         CancellationToken cancellationToken = default);
 }
