@@ -1,6 +1,7 @@
-﻿using System.Threading;
+﻿using System.Collections.Generic;
+using System.Threading;
 using System.Threading.Tasks;
-using DxWorks.ScriptBee.Plugin.Api.ScriptRunner;
+using DxWorks.ScriptBee.Plugin.Api;
 using ScriptBee.Models;
 using ScriptBee.ProjectContext;
 
@@ -8,6 +9,10 @@ namespace ScriptBeeWebApp.Services;
 
 public interface IRunScriptService
 {
+    IScriptRunner? GetScriptRunner(string language);
+    
+    IEnumerable<string> GetSupportedLanguages();
+    
     Task<RunModel?> RunAsync(IScriptRunner scriptRunner, Project project, ProjectModel projectModel,
         string scriptFilePath, CancellationToken cancellationToken = default);
 }
