@@ -1,7 +1,5 @@
 using System;
-using DxWorks.ScriptBee.Plugin.Api.HelperFunctions;
-using DxWorks.ScriptBee.Plugin.Api.ScriptGeneration;
-using DxWorks.ScriptBee.Plugin.Api.ScriptRunner;
+using DxWorks.ScriptBee.Plugin.Api;
 using FluentValidation;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
@@ -51,13 +49,12 @@ public class Startup
         services.AddSingleton<ILogger>(_ => new LoggerConfiguration().CreateLogger());
         services.AddSingleton<ILoadersHolder, LoadersHolder>();
         services.AddSingleton<ILinkersHolder, LinkersHolder>();
-        services.AddSingleton<IScriptGeneratorStrategyHolder, ScriptGeneratorStrategyHolder>();
-        services.AddSingleton<IFileContentProvider, RelativeFileContentProvider>();
+        services.AddSingleton<IPluginRepository, PluginRepository>();
+        services.AddSingleton<IPluginService, PluginService>();
         services.AddSingleton<IProjectManager, ProjectManager>();
         services.AddSingleton<IProjectFileStructureManager, ProjectFileStructureManager>(_ =>
             new ProjectFileStructureManager(userFolderPath));
         services.AddSingleton<IHelperFunctionsFactory, HelperFunctionsFactory>();
-        services.AddSingleton<IHelperFunctionsMapper, HelperFunctionsMapper>();
         services.AddSingleton<IFileNameGenerator, FileNameGenerator>();
         services.AddSingleton<IFileModelService, FileModelService>();
         services.AddSingleton<IProjectStructureService, ProjectStructureService>();
@@ -65,13 +62,22 @@ public class Startup
         services.AddSingleton<IRunModelService, RunModelService>();
         services.AddSingleton<IPluginDiscriminatorHolder, PluginDiscriminatorHolder>();
         services.AddSingleton<IPluginManifestYamlFileReader, PluginManifestYamlFileReader>();
+        services.AddSingleton<IGenerateScriptService, GenerateScriptService>();
         services.AddSingleton<IFileService, FileService>();
         services.AddSingleton<IFileModelService, FileModelService>();
         services.AddSingleton<IPluginManifestValidator, PluginManifestValidator>();
-        services.AddSingleton<IPluginRepository, CompositePluginRepository>();
+        services.AddSingleton<IPluginReader, PluginReader>();
+        services.AddSingleton<IDllLoader, DllLoader>();
         services.AddSingleton<IPluginLoaderFactory, PluginLoaderFactory>();
-        services.AddSingleton<IPluginManifestReader, PluginManifestReader>();
         services.AddSingleton<PluginManager>();
+
+        // add Plugin Loaders
+        // check if services.TryAddEnumerable is needed
+        services.AddSingleton<IPluginLoader, ScriptGeneratorPluginLoader>();
+        services.AddSingleton<IPluginLoader, UiPluginLoader>();
+        services.AddSingleton<IPluginLoader, LinkerPluginLoader>();
+        services.AddSingleton<IPluginLoader, LoaderPluginLoader>();
+
         services.AddValidatorsFromAssemblyContaining<IValidationMarker>();
 
         services.AddControllers();
